Move mistune bits into legacy, sphinx bits into sphinx
